requirements.txt
Jinja2==2.7
MarkupSafe==0.18
Pygments==1.6
WebOb==1.2.3
distribute==0.6.45
gevent==0.13.8
gevent-socketio==0.3.5-rc2
gevent-websocket==0.3.6
greenlet==0.4.1
py==1.4.14
pytest==2.3.5
redis==2.7.5
requests==1.2.3
ws4py-ivideon==0.2.0
wsgiref==0.1.2
